% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationWrapper.R
\name{Correlation}
\alias{Correlation}
\title{Correlation}
\usage{
Correlation(
  data = NULL,
  version = "1",
  alternative = "twoSided",
  assumptionCheckMultivariateEnergy = FALSE,
  assumptionCheckMultivariateMardia = FALSE,
  assumptionCheckMultivariateRoyston = FALSE,
  assumptionCheckMultivariateShapiro = FALSE,
  assumptionCheckPairwiseEnergy = FALSE,
  assumptionCheckPairwiseMardia = FALSE,
  assumptionCheckPairwiseRoyston = FALSE,
  assumptionCheckPairwiseShapiro = FALSE,
  ci = FALSE,
  ciBootstrap = FALSE,
  ciBootstrapSamples = 1000,
  ciLevel = 0.95,
  covariance = FALSE,
  crossProducts = FALSE,
  effectSize = FALSE,
  heatmapPlot = FALSE,
  kendallsTauB = FALSE,
  meansAndSd = FALSE,
  naAction = "pairwise",
  pairwiseDisplay = TRUE,
  partialOutVariables = list(types = list(), value = list()),
  pearson = TRUE,
  plotHeight = 320,
  plotWidth = 480,
  sampleSize = FALSE,
  scatterPlot = FALSE,
  scatterPlotCi = FALSE,
  scatterPlotCiLevel = 0.95,
  scatterPlotDensity = FALSE,
  scatterPlotPredictionInterval = FALSE,
  scatterPlotPredictionIntervalLevel = 0.95,
  scatterPlotStatistic = FALSE,
  significanceFlagged = FALSE,
  significanceReport = TRUE,
  spearman = FALSE,
  variables = list(types = list(), value = list()),
  vovkSellke = FALSE
)
}
\arguments{
\item{ci, }{Confidence intervals for the population correlation (available only for the Pearson correlation). The default is set at 95%, but this percentage can be adjusted. Additionally, you can set up bootstrap estimation, using 1000 replications by default but it can be modified to your desired number.
Defaults to \code{FALSE}.}

\item{covariance, }{The covariance between each pair of variables.
Defaults to \code{FALSE}.}

\item{effectSize, }{The Fisher transformed effect size with standard error.
Defaults to \code{FALSE}.}

\item{heatmapPlot, }{Displays a correlation heatmap for Pearson, Spearman, and Kendall's tau B coefficients separately.
Defaults to \code{FALSE}.}

\item{kendallsTauB, }{Kendall's tau-b rank-order correlation coefficient to quantify the monotonic association between two variables by comparing concordant and non-concordant pairs. Use when the data are not normally distributed.
Defaults to \code{FALSE}.}

\item{pairwiseDisplay, }{Displays a table where one row corresponds to one pair of the specified variables, and the scatter plots are shown individually for each pair. If unticked, the results are presented in matrix format, with variable names in the columns and rows.
Defaults to \code{TRUE}.}

\item{partialOutVariables, }{Variables to partial out in order to compute partial correlations.}

\item{pearson, }{Pearson's product moment correlation coefficient.
Defaults to \code{TRUE}.}

\item{sampleSize, }{The number of complete observations for a given pair of variables.
Defaults to \code{FALSE}.}

\item{scatterPlot, }{Displays scatter plots for all variable pairs. In a matrix format, these are placed above the diagonal. For partial correlations, plots show the relationship between X and Y after removing the effect of Z, and axes then represent residuals from regressing X and Y on Z.
Defaults to \code{FALSE}.}

\item{scatterPlotCi, }{Displays the specified confidence interval around the regression line on the plot.
Defaults to \code{FALSE}.}

\item{scatterPlotDensity, }{Display histogram and the corresponding density plot for each variable. In a matrix format, these are placed on the diagonal
Defaults to \code{FALSE}.}

\item{scatterPlotStatistic, }{Display the correlation coefficient(s) in the plot. This option also adds the x% confidence interval(s) as specified in the Confidence Intervals option.
Defaults to \code{FALSE}.}

\item{significanceFlagged, }{Marks statistically significant correlations.
Defaults to \code{FALSE}.}

\item{significanceReport, }{Displays the p-value corresponding to the observed correlation.
Defaults to \code{TRUE}.}

\item{spearman, }{Spearman's rank-order correlation coefficient to quantify the monotonic association between two variables by ranking the observations. Use when the data are not normally distributed.
Defaults to \code{FALSE}.}

\item{variables, }{Variables for which to compute the correlation coefficient.}

\item{vovkSellke, }{Shows the maximum ratio of the likelihood of the obtained p value under H1 vs the likelihood of the obtained p value under H0. For example, if the two-sided p-value equals .05, the Vovk-Sellke MPR equals 2.46, indicating that this p-value is at most 2.46 times more likely to occur under H1 than under H0.
Defaults to \code{FALSE}.}
}
\description{
The Correlation analysis allows estimation of the population correlation, as well as testing the null hypothesis that the population correlation between pairs of variables equals 0. All possible pairs of the specified variables are analyzed.
## Assumptions (Pearson's rho)
- The variables are both continuous. 
- The data are a random sample from the population. 
- The pairs of variables follow a bivariate normal distribution in the population. 
- The relationship between the variables is linear. 
## Assumptions (Spearman's rho and Kendall's tau).
- Ordinal or continuous variables.
- The data are a random sample from the population.
- The relationship between the pairs of variables is monotonic.
}
