% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationbayesianWrapper.R
\name{CorrelationBayesian}
\alias{CorrelationBayesian}
\title{Bayesian Correlation}
\usage{
CorrelationBayesian(
  data = NULL,
  version = "1",
  alternative = "twoSided",
  bayesFactorReport = TRUE,
  bayesFactorType = "BF10",
  bfRobustnessPlot = FALSE,
  bfRobustnessPlotAdditionalInfo = TRUE,
  bfSequentialPlot = FALSE,
  bfSequentialPlotAdditionalInfo = TRUE,
  ci = FALSE,
  ciLevel = 0.95,
  kendall = FALSE,
  linearityTest = FALSE,
  matrixPlot = FALSE,
  matrixPlotDensity = FALSE,
  matrixPlotPosterior = FALSE,
  naAction = "pairwise",
  pairsMethod = "pearson",
  pairwiseDisplay = TRUE,
  pearson = TRUE,
  plotHeight = 320,
  plotWidth = 480,
  posteriorMedian = FALSE,
  priorPosteriorPlot = FALSE,
  priorPosteriorPlotAdditionalEstimationInfo = TRUE,
  priorPosteriorPlotAdditionalTestingInfo = TRUE,
  priorWidth = 1,
  sampleSize = FALSE,
  scatterPlot = TRUE,
  scatterPlotAdditionalInfo = FALSE,
  seed = 1,
  setSeed = FALSE,
  spearman = FALSE,
  supportCorrelationFlagged = FALSE,
  variablePairs = list(),
  variables = list(types = list(), value = list())
)
}
\arguments{
\item{bayesFactorReport, }{Reports Bayes factor for each test.
Defaults to \code{TRUE}.}

\item{bfRobustnessPlot, }{Displays the Bayes factor as a function of the width of the stretched beta prior on the correlation for each specified pair of variables. The width of the kappa prior is varied between 0 and 2.
Defaults to \code{FALSE}.}

\item{bfRobustnessPlotAdditionalInfo, }{Adds the Bayes factor computed with the user-defined prior and the maximum obtainable Bayes factor.
Defaults to \code{TRUE}.}

\item{bfSequentialPlot, }{Displays the development of the Bayes factor as the data come in using the user-defined prior for each specified pair of variables.
Defaults to \code{FALSE}.}

\item{bfSequentialPlotAdditionalInfo, }{Adds the Bayes factor computed with the user-defined prior; adds a probability wheel depicting the odds of the data under the null vs. alternative hypothesis; shows the decisiveness of the evidence in terms of Jeffreys' (1961) evidence categories.
Defaults to \code{TRUE}.}

\item{ci, }{Displays central 95% credible intervals for the correlation coefficient. The percentage can be changed.
Defaults to \code{FALSE}.}

\item{kendall, }{Kendall's tau-b rank-order correlation coefficient to quantify the monotonic association between two variables by comparing concordant and non-concordant pairs. Use when the data are not normally distributed.
Defaults to \code{FALSE}.}

\item{linearityTest, }{Compute a test for the linearity of the relationship between the two variables by comparing a linear model to a quadratic model.
Defaults to \code{FALSE}.}

\item{matrixPlot, }{Displays a grid of scatterplots for each possible combination of the selected variables. These are placed above the diagonal.
Defaults to \code{FALSE}.}

\item{matrixPlotDensity, }{Displays a histogram and the corresponding density plot for each variable. These are placed on the diagonal.
Defaults to \code{FALSE}.}

\item{matrixPlotPosterior, }{Displays posterior distribution of the correlation coefficient for each possible combination of the selected variables. These are placed below the diagonal.
Defaults to \code{FALSE}.}

\item{pairwiseDisplay, }{Displays a table where one row corresponds to one pair of the specified variables. If unticked, the results are presented in matrix format, with variable names in the columns and rows.
Defaults to \code{TRUE}.}

\item{pearson, }{Pearson's product moment correlation coefficient. This is the famous r value.
Defaults to \code{TRUE}.}

\item{priorPosteriorPlot, }{Displays the prior and posterior distribution of the correlation under the alternative hypothesis for each specified pair of variables.
Defaults to \code{FALSE}.}

\item{priorPosteriorPlotAdditionalEstimationInfo, }{Adds the median and the 95% credible interval of the posterior distribution of the effect size.
Defaults to \code{TRUE}.}

\item{priorPosteriorPlotAdditionalTestingInfo, }{Adds the Bayes factor computed with the user-defined prior; adds a probability wheel depicting the odds of the data under the null vs. alternative hypothesis (Assuming null and alternative had equal probability a priori).
Defaults to \code{TRUE}.}

\item{sampleSize, }{Reports the sample size for each test.
Defaults to \code{FALSE}.}

\item{scatterPlot, }{Displays scatterplots for each specified pair of variables.
Defaults to \code{TRUE}.}

\item{supportCorrelationFlagged, }{Indicates which correlation coefficients yield Bayes factors greater than 10, 30, and 100.
Defaults to \code{FALSE}.}

\item{variables, }{Variables for which to compute the correlation coefficient.}
}
\description{
The Bayesian Correlation analysis allows for the estimation of the population correlation, as well as the comparison of hypotheses. The three comparisons are (i) between the null hypothesis (H0) that the correlation between pairs of variables equals 0, and the alternative hypothesis (H1) that the population correlation takes its value between -1 and 1; (ii) between H0 and the alternative hypothesis (H+) that the population correlation is positive; and (iii) between H0 and the alternative hypothesis (H-) that the population correlation is negative. All possible pairs of the specified variables are analyzed.
## Assumptions (Pearson's rho)
- The variables are both continuous.
- The data are a random sample from the population.
- The pairs of variables follow a bivariate normal distribution in the population.
 - The relationship between the variables is linear.
## Assumptions (Kendall's tau)
- Ordinal or continuous variables.
- The data are a random sample from the population.
- The relationship between the pairs of variables is monotonic.
}
