% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressionlogisticWrapper.R
\name{RegressionLogistic}
\alias{RegressionLogistic}
\title{Logistic Regression}
\usage{
RegressionLogistic(
  data = NULL,
  version = "1",
  accuracy = FALSE,
  auc = FALSE,
  brierScore = FALSE,
  coefficientBootstrap = FALSE,
  coefficientBootstrapSamples = 5000,
  coefficientCi = FALSE,
  coefficientCiAsOddsRatio = TRUE,
  coefficientCiLevel = 0.95,
  coefficientEstimate = TRUE,
  coefficientStandardized = FALSE,
  conditionalEstimatePlot = FALSE,
  conditionalEstimatePlotCi = 0.95,
  conditionalEstimatePlotPoints = FALSE,
  confusionMatrix = FALSE,
  covarianceRatio = FALSE,
  covariates = list(types = list(), value = list()),
  dependent = list(types = list(), value = ""),
  descriptives = FALSE,
  dfbetas = FALSE,
  dffits = FALSE,
  fMeasure = FALSE,
  factors = list(types = list(), value = list()),
  hMeasure = FALSE,
  independentVsPredictedPlot = FALSE,
  independentVsPredictedPlotIncludeInteractions = TRUE,
  independentVsPredictedPlotUseLogit = TRUE,
  interceptTerm = TRUE,
  leverage = FALSE,
  mahalanobis = FALSE,
  method = "enter",
  modelTerms = list(list(components = list(), name = "model0", title = "Model 0"),
    list(components = list(), name = "model1", title = "Model 1")),
  multicollinearity = FALSE,
  nested = TRUE,
  oddsRatio = TRUE,
  plotHeight = 320,
  plotWidth = 480,
  precision = FALSE,
  precisionRecallPlot = FALSE,
  precisionRecallPlotCutoffLabel = FALSE,
  precisionRecallPlotCutoffStep = 0.2,
  predictionsSavedToData = FALSE,
  predictionsSavedToDataColumn = "",
  quadraticTerms = FALSE,
  residualCasewiseDiagnostic = FALSE,
  residualCasewiseDiagnosticCooksDistanceThreshold = 1,
  residualCasewiseDiagnosticType = "outliersOutside",
  residualCasewiseDiagnosticZThreshold = 3,
  residualStatistic = FALSE,
  residualType = "deviance",
  residualVsFittedPlot = FALSE,
  residualVsPredictorPlot = FALSE,
  residualsSavedToData = FALSE,
  residualsSavedToDataColumn = "",
  residualsSavedToDataType = "raw",
  robustSe = FALSE,
  rocPlot = FALSE,
  rocPlotCutoffLabel = FALSE,
  rocPlotCutoffStep = 0.2,
  sensitivity = FALSE,
  specificity = FALSE,
  squaredPearsonResidualVsFittedPlot = FALSE,
  vovkSellke = FALSE,
  weights = list(types = list(), value = "")
)
}
\arguments{
\item{accuracy, }{How often the model's prediction match the actual outcomes.
Defaults to \code{FALSE}.}

\item{auc, }{Area Under the Curve.
Defaults to \code{FALSE}.}

\item{brierScore, }{Another measure of the accuracy of predictions. Based on the mean squared difference between predicted probabilities and the binary outcome. Lower scores mean more accuracy.
Defaults to \code{FALSE}.}

\item{coefficientBootstrap, }{By selecting this option, bootstrapped estimation is applied. By default, the number of replications is set to 5000. This can be changed into the desired number.
Defaults to \code{FALSE}.}

\item{coefficientCi, }{Coverage of the confidence intervals in percentages. The default value is 95%, which can be changed by the user.
Defaults to \code{FALSE}.}

\item{coefficientEstimate, }{Coefficient estimates, standard errors, z-values, and their corresponding p-values.
Defaults to \code{TRUE}.}

\item{coefficientStandardized, }{Standardized estimates represent estimates where the predictors are standardized (X-standardization).
Defaults to \code{FALSE}.}

\item{conditionalEstimatePlot, }{The plots are conditional in the sense that they display the probability of the categorical dependent variable for all levels of the predictor variables given the reference level of all other factors.
Defaults to \code{FALSE}.}

\item{conditionalEstimatePlotCi, }{Displays the specified confidence interval around the regression line on the plot.}

\item{conditionalEstimatePlotPoints, }{Displays individual data points on the plot.
Defaults to \code{FALSE}.}

\item{confusionMatrix, }{The confusion matrix indicates how well the model predicts the outcomes. The table is showing actual versus predicted outcomes and can be used to determine the accuracy of the model.
Defaults to \code{FALSE}.}

\item{covarianceRatio, }{The degree to which a case influences the variance of the regression parameters. Indicated in the table when the covariance ratio is greater than 3 * k/(n-k).
Defaults to \code{FALSE}.}

\item{covariates, }{Continuous predictor variable(s). If ordinal variables are entered it is assumed that their levels are equidistant. Hence, ordinal variables are treated as continuous predictor variables.}

\item{dependent, }{The variable of interest. This is also called the outcome variable. In case of multiple dependent variables, specify the order in which the predictors are entered into the model i.e., hierarchical regression analysis. A block of one or more predictors represents one step in the hierarchy.}

\item{descriptives, }{The levels of the dependent variable/s and the number of observations per level.
Defaults to \code{FALSE}.}

\item{dfbetas, }{The difference between a parameter estimated using all cases and estimated when one case is excluded. Indicated in the table when the absolute value of DFBETAS is greater than 1.
Defaults to \code{FALSE}.}

\item{dffits, }{The difference between the predicted value for a case when the model is estimated including or excluding that case. Indicated in the table when the absolute value of DFFITS is greater than 3 * sqrt(k/(n-k)).
Defaults to \code{FALSE}.}

\item{fMeasure, }{The F-measure is based on the amount of systematic variance divided by the amount of unsystematic variance i.e., mean squares for the model / the residual mean squares.
Defaults to \code{FALSE}.}

\item{factors, }{Categorical predictor variable(s). Ordinal variables here are treated as categorical predictor variables, thus, the ordinal information is ignored.}

\item{hMeasure, }{A more theoretically robust and advanced alternative to the AUC.
Defaults to \code{FALSE}.}

\item{independentVsPredictedPlot, }{Plots the model predictions against each independent variable.
Defaults to \code{FALSE}.}

\item{independentVsPredictedPlotIncludeInteractions, }{Adds every two-way interaction.
Defaults to \code{TRUE}.}

\item{independentVsPredictedPlotUseLogit, }{Plots predicted probabilities on the logit scale, to ensure a linear relation.
Defaults to \code{TRUE}.}

\item{interceptTerm, }{Ticking this box will add a coefficient estimate of the intercept as well. This corresponds to the first level for the independent variable.
Defaults to \code{TRUE}.}

\item{leverage, }{The influence of the observed value of the outcome variable over the predicted values. Indicated in the table when the leverage is greater than 3 * k/n.
Defaults to \code{FALSE}.}

\item{mahalanobis, }{Measures the distance of cases from the mean(s) of the predictor variable(s).
Defaults to \code{FALSE}.}

\item{method, }{Specify the method for entering predictors into the model. For the Enter method, predictors are added as specified in the Model tab. With Forward, Backward, or Stepwise methods, predictors assigned to Model 1 are treated as candidate variables that can be added to or removed from the model based on AIC. Predictors assigned to Model 0 are forced into the model and remain in all steps, regardless of AIC.
\itemize{
  \item \code{"enter"} (default) : All predictors are entered into the models as specified in the Model tab.
  \item \code{"backward"}: All predictors are entered simultaneously, and then removed sequentially based on AIC.
  \item \code{"forward"}: Predictors are entered sequentially based on AIC.
  \item \code{"stepwise"}: Predictors are entered sequentially based on AIC; after each step, the least useful predictor is removed.
}}

\item{modelTerms, }{The independent variables and covariates included in the model. By default, all the main effects and interaction effects of the specified independent variables, and the covariates are included in the model.}

\item{multicollinearity, }{Displays Tolerance and Variance Inflation Factor for each predictor in the model to assess multicollinearity.
Defaults to \code{FALSE}.}

\item{oddsRatio, }{Odds ratio is an indicator of the change in odds resulting from a unit change in the predictor.
Defaults to \code{TRUE}.}

\item{precision, }{Precision describes the proportion of true positives to all positives. Also called the positive predictive value.
Defaults to \code{FALSE}.}

\item{precisionRecallPlot, }{Displays the Positive predicitve value on the y axis vs the true positive rate/sensitivity on the x axis. Useful for imbalanced datasets. The cutoff step determines the threshold at which an observation is classified as positive or negative. 
Defaults to \code{FALSE}.}

\item{precisionRecallPlotCutoffLabel, }{By clicking this box, the specified cutoff step is displayed on the precision recall plot.
Defaults to \code{FALSE}.}

\item{predictionsSavedToData, }{Save the predictions of the most complex model as a new column in the data file
Defaults to \code{FALSE}.}

\item{quadraticTerms, }{Include quadratic terms for the covariates in each model.
Defaults to \code{FALSE}.}

\item{residualCasewiseDiagnostic, }{Provides casewise diagnostics for residuals, displaying cases above a threshold based on standardized residuals (default: 3) or Cook's distance (default: 1), or displaying all cases.
Defaults to \code{FALSE}.}

\item{residualStatistic, }{Displays descriptive statistics of the residuals and predicted values.
Defaults to \code{FALSE}.}

\item{residualVsFittedPlot, }{Scatterplot of the values of the residuals against the predicted values.
Defaults to \code{FALSE}.}

\item{residualVsPredictorPlot, }{Scatterplot for every independent variable and covariate of the residuals and the levels of the variable of interest.
Defaults to \code{FALSE}.}

\item{residualsSavedToData, }{Save the residuals of the most complex model as a new column in the data file.
Defaults to \code{FALSE}.}

\item{residualsSavedToDataType, }{Choose the type of residual to be appended. Raw residuals are simply the differences between observation and model prediction, standardized residuals divide each residual by an estimate of its standard deviation (using rstandard()), whereas studentized residuals divide each residual by an unbiased estimate of its standard deviation computed by refitting the model without that observation  (using rstudent()).
\itemize{
  \item \code{"raw"}
  \item \code{"standard"}
  \item \code{"student"}
}}

\item{robustSe, }{This option controls for errors that are not independent and identically distributed. The use of robust standard errors will not change the coefficient estimates. If this option is not selected the normal standard error will be computed.
Defaults to \code{FALSE}.}

\item{rocPlot, }{Plots performance of the model by plotting the true positive rate to the false positive rate. The cutoff step determines the threshold at which an observation is classified as positive or a negative.
Defaults to \code{FALSE}.}

\item{rocPlotCutoffLabel, }{By clicking this box, the specified cutoff step is displayed on the performance plot.
Defaults to \code{FALSE}.}

\item{sensitivity, }{Sensitivity describes the proportion of true positives.
Defaults to \code{FALSE}.}

\item{specificity, }{Specificity describes the proportion of true negatives.
Defaults to \code{FALSE}.}

\item{squaredPearsonResidualVsFittedPlot, }{With the Squared Pearson residuals plot one can check for overdispersion of the model. Overdispersion indicates that the actual data show greater variability than the model has predicted.
Defaults to \code{FALSE}.}

\item{vovkSellke, }{Shows the maximum ratio of the likelihood of the obtained p value under H1 vs the likelihood of the obtained p value under H0. For example, if the two-sided p-value equals .05, the Vovk-Sellke MPR equals 2.46, indicating that this p-value is at most 2.46 times more likely to occur under H1 than under H0.
Defaults to \code{FALSE}.}

\item{weights, }{The weights used for weighted least squares regression.}
}
\description{
Logistic regression allows the user to model a linear relationship between one or more explanatory variable/s and a categorical dependent variable.
## Assumptions
- The dependent variable is binary.
- The observations are independent of one another.
- The logit (log-odds) of the outcome is linearly related to the continuous predictors.
- There is little or no multicollinearity among the predictors.
}
